pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepllm-scanner"
version = "0.1.0"
description = "CPU-only LLM DAST Framework - OWASP LLM Top 10 Scanner"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DeepLLMScanner Team" }
]
keywords = [
    "llm",
    "security",
    "dast",
    "owasp",
    "scanner",
    "ai-security",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "llama-cpp-python>=0.2.50",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
embedding = [
    "sentence-transformers>=2.2.0",
]

[project.scripts]
deepllm = "deepllm_scanner.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/deepllm_scanner"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
